requirements.txt
boto3==1.28.66
botocore==1.31.66
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.18
